+++ docs/master/sidebar.js
@@ -31,6 +31,7 @@ module.exports = [
         title: "Testing",
         children: [
             'testing/phpunit',
+            'testing/extensions',
         ],
     },
     {
@@ -69,6 +70,7 @@ module.exports = [
             'security/authentication',
             'security/authorization',
             'security/validation',
+            'security/sanitization',
             ['security/resource-exhaustion', 'Resource Exhaustion'],
         ]
     },